cli: refactor ibc and counter into own binaries
