+++ cmd/adam/main.go
@@ -4,20 +4,25 @@ import (
 	"os"
 
 	"github.com/tendermint/basecoin/cmd/commands"
+
 	"github.com/urfave/cli"
 )
 
+func init() {
+	commands.RegisterIBC()
+}
+
 func main() {
 	app := cli.NewApp()
-	app.Name = "basecoin"
-	app.Usage = "basecoin [command] [args...]"
+	app.Name = "adam"
+	app.Usage = "adam [command] [args...]"
 	app.Version = "0.1.0"
 	app.Commands = []cli.Command{
 		commands.StartCmd,
 		commands.TxCmd,
 		commands.QueryCmd,
 		commands.VerifyCmd, // TODO: move to merkleeyes?
-		commands.BlockCmd,  // TODO: move to adam?
+		commands.BlockCmd,
 		commands.AccountCmd,
 	}
 	app.Run(os.Args)
