+++ cmd/commands/ibc.go
@@ -9,6 +9,7 @@ import (
 	"github.com/urfave/cli"
 
 	"github.com/tendermint/basecoin/plugins/ibc"
+	"github.com/tendermint/basecoin/types"
 
 	cmn "github.com/tendermint/go-common"
 	"github.com/tendermint/go-merkle"
@@ -16,25 +17,99 @@ import (
 	tmtypes "github.com/tendermint/tendermint/types"
 )
 
+// Register the IBC plugin at start and for transactions
+func RegisterIBC() {
+	RegisterTxSubcommand(IbcCmd)
+	RegisterStartPlugin("ibc", func() types.Plugin {
+		return ibc.New()
+	})
+}
+
+//---------------------------------------------------------------------
+// ibc flags
+
+var (
+	IbcChainIDFlag = cli.StringFlag{
+		Name:  "chain_id",
+		Usage: "ChainID for the new blockchain",
+		Value: "",
+	}
+
+	IbcGenesisFlag = cli.StringFlag{
+		Name:  "genesis",
+		Usage: "Genesis file for the new blockchain",
+		Value: "",
+	}
+
+	IbcHeaderFlag = cli.StringFlag{
+		Name:  "header",
+		Usage: "Block header for an ibc update",
+		Value: "",
+	}
+
+	IbcCommitFlag = cli.StringFlag{
+		Name:  "commit",
+		Usage: "Block commit for an ibc update",
+		Value: "",
+	}
+
+	IbcFromFlag = cli.StringFlag{
+		Name:  "from",
+		Usage: "Source ChainID",
+		Value: "",
+	}
+
+	IbcToFlag = cli.StringFlag{
+		Name:  "to",
+		Usage: "Destination ChainID",
+		Value: "",
+	}
+
+	IbcTypeFlag = cli.StringFlag{
+		Name:  "type",
+		Usage: "IBC packet type (eg. coin)",
+		Value: "",
+	}
+
+	IbcPayloadFlag = cli.StringFlag{
+		Name:  "payload",
+		Usage: "IBC packet payload",
+		Value: "",
+	}
+
+	IbcPacketFlag = cli.StringFlag{
+		Name:  "packet",
+		Usage: "hex-encoded IBC packet",
+		Value: "",
+	}
+
+	IbcProofFlag = cli.StringFlag{
+		Name:  "proof",
+		Usage: "hex-encoded proof of IBC packet from source chain",
+		Value: "",
+	}
+
+	IbcSequenceFlag = cli.IntFlag{
+		Name:  "sequence",
+		Usage: "sequence number for IBC packet",
+		Value: 0,
+	}
+
+	IbcHeightFlag = cli.IntFlag{
+		Name:  "height",
+		Usage: "Height the packet became egress in source chain",
+		Value: 0,
+	}
+)
+
+//---------------------------------------------------------------------
+// ibc commands
+
 var (
 	IbcCmd = cli.Command{
 		Name:  "ibc",
 		Usage: "Send a transaction to the interblockchain (ibc) plugin",
-		Flags: []cli.Flag{
-			NodeFlag,
-			ChainIDFlag,
-
-			FromFlag,
-
-			AmountFlag,
-			CoinFlag,
-			GasFlag,
-			FeeFlag,
-			SeqFlag,
-
-			NameFlag,
-			DataFlag,
-		},
+		Flags: TxFlags,
 		Subcommands: []cli.Command{
 			IbcRegisterTxCmd,
 			IbcUpdateTxCmd,
@@ -107,6 +182,9 @@ var (
 		},
 	}
 )
+
+//---------------------------------------------------------------------
+// ibc command implementations
 
 func cmdIBCRegisterTx(c *cli.Context) error {
 	chainID := c.String("chain_id")
