+++ cmd/commands/query.go
